# application/src/jreleaser/templates/article/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Ublhub API",
    "description" : "Ublhub REST API definitions for you to consume.",
    "termsOfService" : "https://github.com/project-openubl/ublhub/blob/master/CODE_OF_CONDUCT.md",
    "contact" : {
      "name" : "Ublhub API Support",
      "url" : "https://project-openubl.github.io/",
      "email" : "projectopenubl@googlegroups.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/api/api/projects" : {
      "get" : {
        "tags" : [ "Project Resource" ],
        "summary" : "List projects",
        "description" : "List all projects",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        } ]
      },
      "post" : {
        "tags" : [ "Project Resource" ],
        "summary" : "Create project",
        "description" : "Create a project",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
...
    "/api/api/projects/{projectId}/documents" : {
      "get" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "documentType",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "filterText",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "10",
            "type" : "integer"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "0",
            "type" : "integer"
          }
        }, {
          "name" : "ruc",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "sort_by",
          "in" : "query",
          "schema" : {
            "default" : "created:desc",
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageDtoDocumentDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/api/projects/{projectId}/documents/{documentId}" : {
      "get" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/api/projects/{projectId}/documents/{documentId}/cdr" : {
      "get" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "unzip",
          "in" : "query",
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/api/projects/{projectId}/documents/{documentId}/xml" : {
      "get" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "unzip",
          "in" : "query",
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/api/projects/{projectId}/enrich-document" : {
      "post" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/api/projects/{projectId}/render-document" : {
      "post" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/xml" : {
                "schema" : {
                  "type" : "string"
                }
              },
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/api/projects/{projectId}/upload/document" : {
      "post" : {
        "tags" : [ "Document Resource" ],
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "$ref" : "#/components/schemas/FileUpload"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DocumentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "created" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updated" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "$ref" : "#/components/schemas/Status"
          }
        }
      },
...
      "FileUpload" : {
        "type" : "object"
      },
...
      "JsonObject" : {
        "type" : "array"
      },
...
      "PageDtoDocumentDto" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentDto"
            }
          }
        }
      },
      "ProjectDto" : {
        "required" : [ "name", "sunat" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "maxLength" : 250,
            "type" : "string"
          },
          "sunat" : {
            "$ref" : "#/components/schemas/SunatDto"
          }
        }
      },
...
      "Status" : {
        "type" : "object",
        "properties" : {
          "inProgress" : {
            "type" : "boolean"
          },
          "xmlData" : {
            "$ref" : "#/components/schemas/XMLData"
          },
          "sunat" : {
            "$ref" : "#/components/schemas/Sunat"
          },
          "error" : {
            "$ref" : "#/components/schemas/Error"
          }
        }
      },
...
      "SunatDto" : {
        "required" : [ "facturaUrl", "guiaUrl", "retencionUrl", "username", "password" ],
        "type" : "object",
        "properties" : {
          "facturaUrl" : {
            "type" : "string"
          },
          "guiaUrl" : {
            "type" : "string"
          },
          "retencionUrl" : {
            "type" : "string"
          },
          "username" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          }
        }
      },
      "XMLData" : {
        "type" : "object",
        "properties" : {
          "ruc" : {
            "type" : "string"
          },
          "serieNumero" : {
            "type" : "string"
          },
          "tipoDocumento" : {
            "type" : "string"
          }
        }
      }
    }
  }
}